Rename the active filters component to match the component convention
app/components/courses/ holds four *_component.rb files and this one
foo/view.rb outlier, which arrived unchanged when the component was
promoted out of Find. *_component.rb is the dominant convention (47
files against 20).

Renaming it also stops the component sharing a module path with the
Courses::ActiveFilters::* services under app/services.
