+++ app/components/courses/active_filters_component.html.erb
